Fix all GMloader scripts with new standards

+++ ports/chipndale/Chip n Dale.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
+
 if [ -d "/opt/system/Tools/PortMaster/" ]; then
   controlfolder="/opt/system/Tools/PortMaster"
 elif [ -d "/opt/tools/PortMaster/" ]; then
@@ -8,6 +9,7 @@ else
 fi
 
 source "$controlfolder/control.txt"
+[ -f "/etc/os-release" ] && source "/etc/os-release"
 
 get_controls
 
@@ -16,32 +18,30 @@ $ESUDO chmod 666 /dev/tty1
 printf "\033c" > /dev/tty0
 printf "\033c" > /dev/tty1
 
-GAMEDIR="/$directory/ports/coi"
-if [ -f "/etc/os-release" ]; then
-  source "/etc/os-release"
-fi 
-
-if [ "$OS_NAME" != "JELOS" ]; then
-  export LD_LIBRARY_PATH="/usr/lib:/usr/lib32:/$directory/ports/coi/lib"
-fi
+GAMEDIR="/$directory/ports/chipndale"
+exec > >(tee "$GAMEDIR/log.txt") 2>&1
  
 cd "$GAMEDIR"
+
 export GMLOADER_DEPTH_DISABLE=1
 export GMLOADER_SAVEDIR="$GAMEDIR/gamedata/"
-export SPA_PLUGIN_DIR="/usr/lib32/spa-0.2"
-export PIPEWIRE_MODULE_DIR="/usr/lib32/pipewire-0.3/"
+export LD_LIBRARY_PATH="/usr/lib:/usr/lib32:/$directory/ports/chipndale/lib"
 
+if [ "$OS_NAME" == "JELOS" ]; then
+  export SPA_PLUGIN_DIR="/usr/lib32/spa-0.2"
+  export PIPEWIRE_MODULE_DIR="/usr/lib32/pipewire-0.3/"
+fi
 
-mv gamedata/data.win gamedata/game.droid
+[ -f "./gamedata/data.win" ] && mv gamedata/data.win gamedata/game.droid
+[ -f "./gamedata/game.win" ] && mv gamedata/game.win gamedata/game.droid
 
 $ESUDO chmod 666 /dev/uinput
-$GPTOKEYB "gmloader" &
+$GPTOKEYB "gmloader" -c "chip.gptk" &
 echo "Loading, please wait... " > /dev/tty0
 
 $ESUDO chmod +x "$GAMEDIR/gmloader"
 
-
-./gmloader IssyosWrapper.apk |& tee log.txt /dev/tty0
+./gmloader chip.apk
 
 $ESUDO kill -9 "$(pidof gptokeyb)"
 $ESUDO systemctl restart oga_events &
