+++ ports/ubermoshcollection/Uber/gamedata/OMEGA.sh
@@ -8,6 +8,7 @@ else
 fi
 
 source "$controlfolder/control.txt"
+[ -f "/etc/os-release" ] && source "/etc/os-release"
 
 get_controls
 
@@ -17,24 +18,21 @@ printf "\033c" > /dev/tty0
 printf "\033c" > /dev/tty1
 
 GAMEDIR="/$directory/ports/Uber/gamedata/OMEGA"
-if [ -f "/etc/os-release" ]; then
-  source "/etc/os-release"
-fi 
+exec > >(tee "$GAMEDIR/log.txt") 2>&1
 
-if [ "$OS_NAME" != "JELOS" ]; then
-  export LD_LIBRARY_PATH="/usr/lib:/usr/lib32:/$directory/ports/Uber/gamedata/lib"
-fi
- 
 cd "$GAMEDIR"
-
 
 export GMLOADER_DEPTH_DISABLE=1
 export GMLOADER_SAVEDIR="$GAMEDIR/gamedata/"
-export SPA_PLUGIN_DIR="/usr/lib32/spa-0.2"
-export PIPEWIRE_MODULE_DIR="/usr/lib32/pipewire-0.3/"
+export LD_LIBRARY_PATH="/usr/lib:/usr/lib32:/$directory/ports/Uber/gamedata/lib"
 
+if [ "$OS_NAME" == "JELOS" ]; then
+  export SPA_PLUGIN_DIR="/usr/lib32/spa-0.2"
+  export PIPEWIRE_MODULE_DIR="/usr/lib32/pipewire-0.3/"
+fi
 
-mv gamedata/data.win gamedata/game.droid
+[ -f "./gamedata/data.win" ] && mv gamedata/data.win gamedata/game.droid
+[ -f "./gamedata/game.win" ] && mv gamedata/game.win gamedata/game.droid
 
 $ESUDO chmod 666 /dev/uinput
 $GPTOKEYB "gmloader" -c "OMEGA.gptk" &
@@ -42,8 +40,7 @@ echo "Loading, please wait... " > /dev/tty0
 
 $ESUDO chmod +x "$GAMEDIR/gmloader"
 
-
-./gmloader uberwrapper3.apk |& tee log.txt /dev/tty0
+./gmloader uberwrapper3.apk
 
 $ESUDO kill -9 "$(pidof gptokeyb)"
 $ESUDO systemctl restart oga_events &
